Generalized ProfileUser to Profile page
