+++ src/pages/index.js
@@ -1,11 +1,13 @@
 import FrontPage from "./FrontPage";
 import LogIn from "./LogIn";
 import LogOut from "./LogOut";
-import SignUp from './SignUp'
+import SignUp from './SignUp';
+import Profile from "./Profile";
 
 export {
     FrontPage,
     LogIn,
     LogOut,
-    SignUp
+    SignUp,
+    Profile
 }
